Move dialogue code of teachers, guild joiners and oracle to UIHouses
