+++ src/Engine/mm7_data.cpp
@@ -2393,36 +2393,6 @@ IndexedArray<IndexedArray<CharacterSkillMastery, CHARACTER_SKILL_FIRST, CHARACTE
     }}
 }};
 
-std::array<std::pair<int16_t, ITEM_TYPE>, 27> _4F0882_evt_VAR_PlayerItemInHands_vals = {{
-    {0x0D4, ITEM_QUEST_VASE},
-    {0x0D5, ITEM_RARE_LADY_CARMINES_DAGGER},
-    {0x0D6, ITEM_MESSAGE_SCROLL_OF_WAVES},
-    {0x0D7, ITEM_MESSAGE_CIPHER},
-    {0x0D8, ITEM_QUEST_WORN_BELT},
-    {0x0D9, ITEM_QUEST_HEART_OF_THE_WOOD},
-    {0x0DA, ITEM_MESSAGE_MAP_TO_EVENMORN_ISLAND},
-    {0x0DB, ITEM_QUEST_GOLEM_HEAD},
-    {0x0DC, ITEM_QUEST_ABBEY_NORMAL_GOLEM_HEAD},
-    {0x0DD, ITEM_QUEST_GOLEM_RIGHT_ARM},
-    {0x0DE, ITEM_QUEST_GOLEM_LEFT_ARM},
-    {0x0DF, ITEM_QUEST_GOLEM_RIGHT_LEG},
-    {0x0E0, ITEM_QUEST_GOLEM_LEFT_LEG},
-    {0x0E1, ITEM_QUEST_GOLEM_CHEST},
-    {0x0E2, ITEM_SPELLBOOK_DIVINE_INTERVENTION},
-    {0x0E3, ITEM_QUEST_DRAGON_EGG},
-    {0x0E4, ITEM_QUEST_ZOKARR_IVS_SKULL},
-    {0x0E5, ITEM_QUEST_LICH_JAR_EMPTY},
-    {0x0E6, ITEM_QUEST_ELIXIR},
-    {0x0E7, ITEM_QUEST_CASE_OF_SOUL_JARS},
-    {0x0E8, ITEM_QUEST_ALTAR_PIECE_1},
-    {0x0E9, ITEM_QUEST_ALTAR_PIECE_2},
-    {0x0EA, ITEM_QUEST_CONTROL_CUBE},
-    {0x0EB, ITEM_QUEST_WETSUIT},
-    {0x0EC, ITEM_QUEST_OSCILLATION_OVERTHRUSTER},
-    {0x0ED, ITEM_QUEST_LICH_JAR_FULL},
-    {0x0F1, ITEM_RARE_THE_PERFECT_BOW}
-}};
-
 int ai_arrays_size;
 std::array<int, 500> ai_near_actors_targets_pid;
 std::array<unsigned int, 500> ai_near_actors_ids;
