+++ src/GUI/GUIWindow.cpp
@@ -87,155 +87,10 @@ GraphicsImage *ui_leather_mm6 = nullptr;
 GraphicsImage *ui_leather_mm7 = nullptr;
 
 DIALOGUE_TYPE _dword_F8B1D8_last_npc_topic_menu;
-AwardType dword_F8B1AC_award_bit_number;
-CharacterSkillType dword_F8B1AC_skill_being_taught; // Address the same as above --- splitting a union into two variables.
 
 std::array<int, 28> possibleAddressingAwardBits = {{1,  2,  3,  4,  5,  7,  32, 33, 36, 37,
                                                     38, 40, 41, 42, 43, 45, 46, 47, 48, 49,
                                                     50, 51, 52, 53, 54, 55, 56, 60}};
-
-// TODO(Nik-RE-dev): move somewhere along with guild join dialogue processing
-IndexedArray<int, GUILD_FIRST, GUILD_LAST> priceForMembership = {{
-    {GUILD_OF_ELEMENTS, 100},
-    {GUILD_OF_SELF,     100},
-    {GUILD_OF_AIR,      50},
-    {GUILD_OF_EARTH,    50},
-    {GUILD_OF_FIRE,     50},
-    {GUILD_OF_WATER,    50},
-    {GUILD_OF_BODY,     50},
-    {GUILD_OF_MIND,     50},
-    {GUILD_OF_SPIRIT,   50},
-    {GUILD_OF_LIGHT,    1000},
-    {GUILD_OF_DARK,     1000}
-}};
-
-// TODO(Nik-RE-dev): move somewhere along with mastery teacing dialogue processing
-IndexedArray<int, CHARACTER_SKILL_FIRST, CHARACTER_SKILL_LAST> expertSkillMasteryCost = {{
-    {CHARACTER_SKILL_STAFF,        2000},
-    {CHARACTER_SKILL_SWORD,        2000},
-    {CHARACTER_SKILL_DAGGER,       2000},
-    {CHARACTER_SKILL_AXE,          2000},
-    {CHARACTER_SKILL_SPEAR,        2000},
-    {CHARACTER_SKILL_BOW,          2000},
-    {CHARACTER_SKILL_MACE,         2000},
-    {CHARACTER_SKILL_BLASTER,      0},
-    {CHARACTER_SKILL_SHIELD,       1000},
-    {CHARACTER_SKILL_LEATHER,      1000},
-    {CHARACTER_SKILL_CHAIN,        1000},
-    {CHARACTER_SKILL_PLATE,        1000},
-    {CHARACTER_SKILL_FIRE,         1000},
-    {CHARACTER_SKILL_AIR,          1000},
-    {CHARACTER_SKILL_WATER,        1000},
-    {CHARACTER_SKILL_EARTH,        1000},
-    {CHARACTER_SKILL_SPIRIT,       1000},
-    {CHARACTER_SKILL_MIND,         1000},
-    {CHARACTER_SKILL_BODY,         1000},
-    {CHARACTER_SKILL_LIGHT,        2000},
-    {CHARACTER_SKILL_DARK,         2000},
-    {CHARACTER_SKILL_ITEM_ID,      500},
-    {CHARACTER_SKILL_MERCHANT,     2000},
-    {CHARACTER_SKILL_REPAIR,       500},
-    {CHARACTER_SKILL_BODYBUILDING, 500},
-    {CHARACTER_SKILL_MEDITATION,   500},
-    {CHARACTER_SKILL_PERCEPTION,   500},
-    {CHARACTER_SKILL_DIPLOMACY,    0}, // not used
-    {CHARACTER_SKILL_THIEVERY,     0}, // not used
-    {CHARACTER_SKILL_TRAP_DISARM,  500},
-    {CHARACTER_SKILL_DODGE,        2000},
-    {CHARACTER_SKILL_UNARMED,      2000},
-    {CHARACTER_SKILL_MONSTER_ID,   500},
-    {CHARACTER_SKILL_ARMSMASTER,   2000},
-    {CHARACTER_SKILL_STEALING,     500},
-    {CHARACTER_SKILL_ALCHEMY,      500},
-    {CHARACTER_SKILL_LEARNING,     2000},
-    {CHARACTER_SKILL_CLUB,         500},
-    {CHARACTER_SKILL_MISC,         0} // hidden, not used
-}};
-
-IndexedArray<int, CHARACTER_SKILL_FIRST, CHARACTER_SKILL_LAST> masterSkillMasteryCost = {{
-    {CHARACTER_SKILL_STAFF,        5000},
-    {CHARACTER_SKILL_SWORD,        5000},
-    {CHARACTER_SKILL_DAGGER,       5000},
-    {CHARACTER_SKILL_AXE,          5000},
-    {CHARACTER_SKILL_SPEAR,        5000},
-    {CHARACTER_SKILL_BOW,          5000},
-    {CHARACTER_SKILL_MACE,         5000},
-    {CHARACTER_SKILL_BLASTER,      0},
-    {CHARACTER_SKILL_SHIELD,       3000},
-    {CHARACTER_SKILL_LEATHER,      3000},
-    {CHARACTER_SKILL_CHAIN,        3000},
-    {CHARACTER_SKILL_PLATE,        3000},
-    {CHARACTER_SKILL_FIRE,         4000},
-    {CHARACTER_SKILL_AIR,          4000},
-    {CHARACTER_SKILL_WATER,        4000},
-    {CHARACTER_SKILL_EARTH,        4000},
-    {CHARACTER_SKILL_SPIRIT,       4000},
-    {CHARACTER_SKILL_MIND,         4000},
-    {CHARACTER_SKILL_BODY,         4000},
-    {CHARACTER_SKILL_LIGHT,        5000},
-    {CHARACTER_SKILL_DARK,         5000},
-    {CHARACTER_SKILL_ITEM_ID,      2500},
-    {CHARACTER_SKILL_MERCHANT,     5000},
-    {CHARACTER_SKILL_REPAIR,       2500},
-    {CHARACTER_SKILL_BODYBUILDING, 2500},
-    {CHARACTER_SKILL_MEDITATION,   2500},
-    {CHARACTER_SKILL_PERCEPTION,   2500},
-    {CHARACTER_SKILL_DIPLOMACY,    0}, // not used
-    {CHARACTER_SKILL_THIEVERY,     0}, // not used
-    {CHARACTER_SKILL_TRAP_DISARM,  2500},
-    {CHARACTER_SKILL_DODGE,        5000},
-    {CHARACTER_SKILL_UNARMED,      5000},
-    {CHARACTER_SKILL_MONSTER_ID,   2500},
-    {CHARACTER_SKILL_ARMSMASTER,   5000},
-    {CHARACTER_SKILL_STEALING,     2500},
-    {CHARACTER_SKILL_ALCHEMY,      2500},
-    {CHARACTER_SKILL_LEARNING,     5000},
-    {CHARACTER_SKILL_CLUB,         2500},
-    {CHARACTER_SKILL_MISC,         0} // hidden, not used
-}};
-
-IndexedArray<int, CHARACTER_SKILL_FIRST, CHARACTER_SKILL_LAST> grandmasterSkillMasteryCost = {{
-    {CHARACTER_SKILL_STAFF,        8000},
-    {CHARACTER_SKILL_SWORD,        8000},
-    {CHARACTER_SKILL_DAGGER,       8000},
-    {CHARACTER_SKILL_AXE,          8000},
-    {CHARACTER_SKILL_SPEAR,        8000},
-    {CHARACTER_SKILL_BOW,          8000},
-    {CHARACTER_SKILL_MACE,         8000},
-    {CHARACTER_SKILL_BLASTER,      0},
-    {CHARACTER_SKILL_SHIELD,       7000},
-    {CHARACTER_SKILL_LEATHER,      7000},
-    {CHARACTER_SKILL_CHAIN,        7000},
-    {CHARACTER_SKILL_PLATE,        7000},
-    {CHARACTER_SKILL_FIRE,         8000},
-    {CHARACTER_SKILL_AIR,          8000},
-    {CHARACTER_SKILL_WATER,        8000},
-    {CHARACTER_SKILL_EARTH,        8000},
-    {CHARACTER_SKILL_SPIRIT,       8000},
-    {CHARACTER_SKILL_MIND,         8000},
-    {CHARACTER_SKILL_BODY,         8000},
-    {CHARACTER_SKILL_LIGHT,        8000},
-    {CHARACTER_SKILL_DARK,         8000},
-    {CHARACTER_SKILL_ITEM_ID,      6000},
-    {CHARACTER_SKILL_MERCHANT,     8000},
-    {CHARACTER_SKILL_REPAIR,       6000},
-    {CHARACTER_SKILL_BODYBUILDING, 6000},
-    {CHARACTER_SKILL_MEDITATION,   6000},
-    {CHARACTER_SKILL_PERCEPTION,   6000},
-    {CHARACTER_SKILL_DIPLOMACY,    0}, // not used
-    {CHARACTER_SKILL_THIEVERY,     0}, // not used
-    {CHARACTER_SKILL_TRAP_DISARM,  6000},
-    {CHARACTER_SKILL_DODGE,        8000},
-    {CHARACTER_SKILL_UNARMED,      8000},
-    {CHARACTER_SKILL_MONSTER_ID,   6000},
-    {CHARACTER_SKILL_ARMSMASTER,   8000},
-    {CHARACTER_SKILL_STEALING,     6000},
-    {CHARACTER_SKILL_ALCHEMY,      6000},
-    {CHARACTER_SKILL_LEARNING,     8000},
-    {CHARACTER_SKILL_CLUB,         6000},
-    {CHARACTER_SKILL_MISC,         0} // hidden, not used
-}};
-
 void SetCurrentMenuID(MENU_STATE uMenu) {
     sCurrentMenuID = uMenu;
     logger->warning("CurrentMenu = {} \n", toString(uMenu));
@@ -570,23 +425,6 @@ GUIWindow::GUIWindow(WindowType windowType, Pointi position, Sizei dimensions, W
     this->sHint = hint;
 
     this->receives_keyboard_input = false;
-}
-
-//----- (004B3EF0) --------------------------------------------------------
-void DrawJoinGuildWindow(GUILD_ID guild_id) {
-    uDialogueType = DIALOGUE_81_join_guild;
-    current_npc_text = pNPCTopics[guild_id + 99].pText;
-    GetJoinGuildDialogueOption(guild_id);
-    pDialogueWindow->Release();
-    pDialogueWindow = new GUIWindow(WINDOW_Dialogue, {0, 0}, {render->GetRenderDimensions().w, 350}, guild_id);
-    pBtn_ExitCancel = pDialogueWindow->CreateButton({471, 445}, {169, 35}, 1, 0, UIMSG_Escape, 0, Io::InputAction::Invalid,
-        localization->GetString(LSTR_CANCEL), { ui_exit_cancel_button_background }
-    );
-    pDialogueWindow->CreateButton({0, 0}, {0, 0}, 1, 0, UIMSG_HouseScreenClick, 0, Io::InputAction::Invalid, "");
-    pDialogueWindow->CreateButton({480, 160}, {140, 30}, 1, 0, UIMSG_ClickNPCTopic, DIALOGUE_82_join_guild, Io::InputAction::Invalid,
-        localization->GetString(LSTR_JOIN));
-    pDialogueWindow->_41D08F_set_keyboard_control_group(1, 1, 0, 2);
-    dialog_menu_id = DIALOGUE_OTHER;
 }
 
 void DialogueEnding() {
@@ -1225,203 +1063,6 @@ void ClickNPCTopic(DIALOGUE_TYPE topic) {
     BackToHouseMenu();
 }
 
-
-void _4B3FE5_training_dialogue(int eventId) {
-    uDialogueType = DIALOGUE_SKILL_TRAINER;
-    current_npc_text = std::string(pNPCTopics[eventId + 168].pText);
-    _4B254D_SkillMasteryTeacher(eventId);  // checks whether the facility can be used
-    pDialogueWindow->Release();
-    pDialogueWindow = new GUIWindow(WINDOW_Dialogue, {0, 0}, {render->GetRenderDimensions().w, 350}, eventId);
-    pBtn_ExitCancel = pDialogueWindow->CreateButton({471, 445}, {169, 35}, 1, 0, UIMSG_Escape, 0, Io::InputAction::Invalid,
-        localization->GetString(LSTR_CANCEL), { ui_exit_cancel_button_background }
-    );
-    pDialogueWindow->CreateButton({0, 0}, {0, 0}, 1, 0, UIMSG_HouseScreenClick, 0, Io::InputAction::Invalid, "");
-    pDialogueWindow->CreateButton({480, 160}, {0x8Cu, 0x1Eu}, 1, 0, UIMSG_ClickNPCTopic, DIALOGUE_79_mastery_teacher, Io::InputAction::Invalid,
-        guild_membership_approved ? localization->GetString(LSTR_LEARN) : "");
-    pDialogueWindow->_41D08F_set_keyboard_control_group(1, 1, 0, 2);
-    dialog_menu_id = DIALOGUE_OTHER;
-}
-
-/**
- * @offset 0x004B1ECE.
- *
- * @brief Oracle's 'I lost it!' dialog option
- */
-void OracleDialogue() {
-    ItemGen *item = nullptr;
-    ITEM_TYPE item_id = ITEM_NULL;
-
-    // display "You never had it" if nothing missing will be found
-    current_npc_text = pNPCTopics[667].pText;
-
-    // only items with special subquest in range 212-237 and also 241 are recoverable
-    for (auto pair : _4F0882_evt_VAR_PlayerItemInHands_vals) {
-        int quest_id = pair.first;
-        if (pParty->_questBits[quest_id]) {
-            ITEM_TYPE search_item_id = pair.second;
-            if (!pParty->hasItem(search_item_id) && pParty->pPickedItem.uItemID != search_item_id) {
-                item_id = search_item_id;
-                break;
-            }
-        }
-    }
-
-    // missing item found
-    if (item_id != ITEM_NULL) {
-        pParty->pCharacters[0].AddVariable(VAR_PlayerItemInHands, std::to_underlying(item_id));
-        // TODO(captainurist): what if fmt throws?
-        current_npc_text = fmt::sprintf(pNPCTopics[666].pText, // "Here's %s that you lost. Be careful"
-                                        fmt::format("{::}{}\f00000", colorTable.Jonquil.tag(),
-                                                    pItemTable->pItems[item_id].pUnidentifiedName));
-    }
-
-    // missing item is lich jar and we need to bind soul vessel to lich class character
-    // TODO(Nik-RE-dev): this code is walking only through inventory, but item was added to hand, so it will not bind new item if it was acquired
-    //                   rather this code will bind jars that already present in inventory to liches that currently do not have binded jars
-    if (item_id == ITEM_QUEST_LICH_JAR_FULL) {
-        for (int i = 0; i < pParty->pCharacters.size(); i++) {
-            if (pParty->pCharacters[i].classType == CHARACTER_CLASS_LICH) {
-                bool have_vessels_soul = false;
-                for (Character &player : pParty->pCharacters) {
-                    for (int idx = 0; idx < Character::INVENTORY_SLOT_COUNT; idx++) {
-                        if (player.pInventoryItemList[idx].uItemID == ITEM_QUEST_LICH_JAR_FULL) {
-                            if (player.pInventoryItemList[idx].uHolderPlayer == -1) {
-                                item = &player.pInventoryItemList[idx];
-                            }
-                            if (player.pInventoryItemList[idx].uHolderPlayer == i) {
-                                have_vessels_soul = true;
-                            }
-                        }
-                    }
-                }
-
-                if (item && !have_vessels_soul) {
-                    item->uHolderPlayer = i;
-                    break;
-                }
-            }
-        }
-    }
-}
-
-//----- (004B254D) --------------------------------------------------------
-std::string _4B254D_SkillMasteryTeacher(int trainerInfo) {
-    uint8_t teacherLevel = (trainerInfo - 200) % 3;
-    CharacterSkillType skillBeingTaught = static_cast<CharacterSkillType>((trainerInfo - 200) / 3);
-    Character *activePlayer = &pParty->activeCharacter();
-    CharacterClassType pClassType = activePlayer->classType;
-    CharacterSkillMastery currClassMaxMastery = skillMaxMasteryPerClass[pClassType][skillBeingTaught];
-    CharacterSkillMastery masteryLevelBeingTaught = dword_F8B1B0_MasteryBeingTaught = static_cast<CharacterSkillMastery>(teacherLevel + 2);
-    guild_membership_approved = false;
-
-    if (currClassMaxMastery < masteryLevelBeingTaught) {
-        if (skillMaxMasteryPerClass[getTier2Class(pClassType)][skillBeingTaught] >= masteryLevelBeingTaught) {
-            return localization->FormatString(LSTR_FMT_HAVE_TO_BE_PROMOTED, localization->GetClassName(getTier2Class(pClassType)));
-        } else if (skillMaxMasteryPerClass[getTier3LightClass(pClassType)][skillBeingTaught] >= masteryLevelBeingTaught &&
-                skillMaxMasteryPerClass[getTier3DarkClass(pClassType)][skillBeingTaught] >= masteryLevelBeingTaught) {
-            return localization->FormatString(LSTR_FMT_HAVE_TO_BE_PROMOTED_2,
-                    localization->GetClassName(getTier3LightClass(pClassType)),
-                    localization->GetClassName(getTier3DarkClass(pClassType)));
-        } else if (skillMaxMasteryPerClass[getTier3LightClass(pClassType)][skillBeingTaught] >= masteryLevelBeingTaught) {
-            return localization->FormatString(LSTR_FMT_HAVE_TO_BE_PROMOTED, localization->GetClassName(getTier3LightClass(pClassType)));
-        } else if (skillMaxMasteryPerClass[getTier3DarkClass(pClassType)][skillBeingTaught] >= masteryLevelBeingTaught) {
-            return localization->FormatString(LSTR_FMT_HAVE_TO_BE_PROMOTED, localization->GetClassName(getTier3DarkClass(pClassType)));
-        } else {
-            return localization->FormatString(LSTR_FMT_SKILL_CANT_BE_LEARNED, localization->GetClassName(pClassType));
-        }
-    }
-
-    // Not in your condition!
-    if (!activePlayer->CanAct()) {
-        return std::string(pNPCTopics[122].pText);
-    }
-
-    // You must know the skill before you can become an expert in it!
-    int skillLevel = activePlayer->getSkillValue(skillBeingTaught).level();
-    if (!skillLevel) {
-        return std::string(pNPCTopics[131].pText);
-    }
-
-    // You are already have this mastery in this skill.
-    CharacterSkillMastery skillMastery = activePlayer->getSkillValue(skillBeingTaught).mastery();
-    if (std::to_underlying(skillMastery) > teacherLevel + 1) {
-        return std::string(pNPCTopics[teacherLevel + 128].pText);
-    }
-
-    dword_F8B1AC_skill_being_taught = skillBeingTaught;
-
-    bool canLearn = true;
-
-    if (masteryLevelBeingTaught == CHARACTER_SKILL_MASTERY_EXPERT) {
-        canLearn = skillLevel >= 4;
-        gold_transaction_amount = expertSkillMasteryCost[skillBeingTaught];
-    }
-
-    if (masteryLevelBeingTaught == CHARACTER_SKILL_MASTERY_MASTER) {
-        switch (skillBeingTaught) {
-          case CHARACTER_SKILL_LIGHT:
-            canLearn = pParty->_questBits[114];
-            break;
-          case CHARACTER_SKILL_DARK:
-            canLearn = pParty->_questBits[110];
-            break;
-          case CHARACTER_SKILL_MERCHANT:
-            canLearn = activePlayer->GetBasePersonality() >= 50;
-            break;
-          case CHARACTER_SKILL_BODYBUILDING:
-            canLearn = activePlayer->GetBaseEndurance() >= 50;
-            break;
-          case CHARACTER_SKILL_LEARNING:
-            canLearn = activePlayer->GetBaseIntelligence() >= 50;
-            break;
-          default:
-            break;
-        }
-        canLearn = canLearn && (skillLevel >= 7);
-        gold_transaction_amount = masterSkillMasteryCost[skillBeingTaught];
-    }
-
-    if (masteryLevelBeingTaught == CHARACTER_SKILL_MASTERY_GRANDMASTER) {
-        switch (skillBeingTaught) {
-          case CHARACTER_SKILL_LIGHT:
-            canLearn = activePlayer->isClass(CHARACTER_CLASS_ARCHAMGE) || activePlayer->isClass(CHARACTER_CLASS_PRIEST_OF_SUN);
-            break;
-          case CHARACTER_SKILL_DARK:
-            canLearn = activePlayer->isClass(CHARACTER_CLASS_LICH) || activePlayer->isClass(CHARACTER_CLASS_PRIEST_OF_MOON);
-            break;
-          case CHARACTER_SKILL_DODGE:
-            canLearn = activePlayer->pActiveSkills[CHARACTER_SKILL_UNARMED].level() >= 10;
-            break;
-          case CHARACTER_SKILL_UNARMED:
-            canLearn = activePlayer->pActiveSkills[CHARACTER_SKILL_DODGE].level() >= 10;
-            break;
-          default:
-            break;
-        }
-        canLearn = canLearn && (skillLevel >= 10);
-        gold_transaction_amount = grandmasterSkillMasteryCost[skillBeingTaught];
-    }
-
-    // You don't meet the requirements, and cannot be taught until you do.
-    if (!canLearn) {
-        return std::string(pNPCTopics[127].pText);
-    }
-
-    // You don't have enough gold!
-    if (gold_transaction_amount > pParty->GetGold()) {
-        return std::string(pNPCTopics[124].pText);
-    }
-
-    guild_membership_approved = true;
-
-    return localization->FormatString(
-        LSTR_FMT_BECOME_S_IN_S_FOR_D_GOLD,
-        localization->MasteryNameLong(masteryLevelBeingTaught),
-        localization->GetSkillName(skillBeingTaught),
-        gold_transaction_amount
-    );
-}
-
 std::string BuildDialogueString(const char *lpsz, uint8_t uPlayerID, ItemGen *a3, HOUSE_ID houseId, int a5, GameTime *a6) {
     std::string str = std::string(lpsz);
     return BuildDialogueString(str, uPlayerID, a3, houseId, a5, a6);
@@ -1899,31 +1540,4 @@ std::string GetDisplayName(Actor *actor) {
         return pMonsterStats->pPlaceStrings[actor->dword_000334_unique_name];
     else
         return pMonsterStats->pInfos[actor->monsterInfo.uID].pName;
-}
-
-//----- (004B29F2) --------------------------------------------------------
-const std::string &GetJoinGuildDialogueOption(GUILD_ID guild_id) {
-    static const int dialogue_base = 110;
-    guild_membership_approved = false;
-    dword_F8B1AC_award_bit_number = static_cast<AwardType>(Award_Membership_ElementalGuilds + std::to_underlying(guild_id));
-    gold_transaction_amount = priceForMembership[guild_id];
-
-    // TODO(pskelton): check this behaviour
-    if (!pParty->hasActiveCharacter())
-        pParty->setActiveToFirstCanAct();  // avoid nzi
-
-    if (pParty->activeCharacter().CanAct()) {
-        if (pParty->activeCharacter()._achievedAwardsBits[dword_F8B1AC_award_bit_number]) {
-            return pNPCTopics[dialogue_base + 13].pText;
-        } else {
-            if (gold_transaction_amount <= pParty->GetGold()) {
-                guild_membership_approved = true;
-                return pNPCTopics[dialogue_base + guild_id].pText;
-            } else {
-                return pNPCTopics[dialogue_base + 14].pText;
-            }
-        }
-    } else {
-        return pNPCTopics[dialogue_base + 12].pText;
-    }
 }
