+++ src/GUI/UI/UIHouses.cpp
@@ -2,6 +2,7 @@
 
 #include <cstdlib>
 #include <vector>
+#include <utility>
 
 #include "Engine/AssetsManager.h"
 #include "Engine/Engine.h"
@@ -19,6 +20,7 @@
 #include "Engine/Tables/BuildingTable.h"
 #include "Engine/Tables/AwardTable.h"
 #include "Engine/Tables/TransitionTable.h"
+#include "Engine/Tables/ItemTable.h"
 #include "Engine/Objects/NPC.h"
 
 #include "GUI/GUIButton.h"
@@ -55,6 +57,10 @@ BuildingType in_current_building_type;  // 00F8B198
 DIALOGUE_TYPE dialog_menu_id;     // 00F8B19C
 
 GraphicsImage *_591428_endcap = nullptr;
+
+// TODO(Nik-RE-dev): refactor and remove
+AwardType dword_F8B1AC_award_bit_number;
+CharacterSkillType dword_F8B1AC_skill_being_taught; // Address the same as above --- splitting a union into two variables.
 
 std::vector<HouseNpcDesc> houseNpcs;
 int currentHouseNpc;
@@ -257,6 +263,176 @@ std::array<const HouseAnimDescr, 196> pAnimatedRooms = { {  // 0x4E5F70
     { "Player Castle Good", 0x24, 0, BUILDING_CASTLE, 0, 0 },
     { "Player Castle Bad", 0x24, 0, BUILDING_CASTLE, 0, 0 }
 } };
+
+IndexedArray<int, GUILD_FIRST, GUILD_LAST> priceForMembership = {{
+    {GUILD_OF_ELEMENTS, 100},
+    {GUILD_OF_SELF,     100},
+    {GUILD_OF_AIR,      50},
+    {GUILD_OF_EARTH,    50},
+    {GUILD_OF_FIRE,     50},
+    {GUILD_OF_WATER,    50},
+    {GUILD_OF_BODY,     50},
+    {GUILD_OF_MIND,     50},
+    {GUILD_OF_SPIRIT,   50},
+    {GUILD_OF_LIGHT,    1000},
+    {GUILD_OF_DARK,     1000}
+}};
+
+IndexedArray<int, CHARACTER_SKILL_FIRST, CHARACTER_SKILL_LAST> expertSkillMasteryCost = {{
+    {CHARACTER_SKILL_STAFF,        2000},
+    {CHARACTER_SKILL_SWORD,        2000},
+    {CHARACTER_SKILL_DAGGER,       2000},
+    {CHARACTER_SKILL_AXE,          2000},
+    {CHARACTER_SKILL_SPEAR,        2000},
+    {CHARACTER_SKILL_BOW,          2000},
+    {CHARACTER_SKILL_MACE,         2000},
+    {CHARACTER_SKILL_BLASTER,      0},
+    {CHARACTER_SKILL_SHIELD,       1000},
+    {CHARACTER_SKILL_LEATHER,      1000},
+    {CHARACTER_SKILL_CHAIN,        1000},
+    {CHARACTER_SKILL_PLATE,        1000},
+    {CHARACTER_SKILL_FIRE,         1000},
+    {CHARACTER_SKILL_AIR,          1000},
+    {CHARACTER_SKILL_WATER,        1000},
+    {CHARACTER_SKILL_EARTH,        1000},
+    {CHARACTER_SKILL_SPIRIT,       1000},
+    {CHARACTER_SKILL_MIND,         1000},
+    {CHARACTER_SKILL_BODY,         1000},
+    {CHARACTER_SKILL_LIGHT,        2000},
+    {CHARACTER_SKILL_DARK,         2000},
+    {CHARACTER_SKILL_ITEM_ID,      500},
+    {CHARACTER_SKILL_MERCHANT,     2000},
+    {CHARACTER_SKILL_REPAIR,       500},
+    {CHARACTER_SKILL_BODYBUILDING, 500},
+    {CHARACTER_SKILL_MEDITATION,   500},
+    {CHARACTER_SKILL_PERCEPTION,   500},
+    {CHARACTER_SKILL_DIPLOMACY,    0}, // not used
+    {CHARACTER_SKILL_THIEVERY,     0}, // not used
+    {CHARACTER_SKILL_TRAP_DISARM,  500},
+    {CHARACTER_SKILL_DODGE,        2000},
+    {CHARACTER_SKILL_UNARMED,      2000},
+    {CHARACTER_SKILL_MONSTER_ID,   500},
+    {CHARACTER_SKILL_ARMSMASTER,   2000},
+    {CHARACTER_SKILL_STEALING,     500},
+    {CHARACTER_SKILL_ALCHEMY,      500},
+    {CHARACTER_SKILL_LEARNING,     2000},
+    {CHARACTER_SKILL_CLUB,         500},
+    {CHARACTER_SKILL_MISC,         0} // hidden, not used
+}};
+
+IndexedArray<int, CHARACTER_SKILL_FIRST, CHARACTER_SKILL_LAST> masterSkillMasteryCost = {{
+    {CHARACTER_SKILL_STAFF,        5000},
+    {CHARACTER_SKILL_SWORD,        5000},
+    {CHARACTER_SKILL_DAGGER,       5000},
+    {CHARACTER_SKILL_AXE,          5000},
+    {CHARACTER_SKILL_SPEAR,        5000},
+    {CHARACTER_SKILL_BOW,          5000},
+    {CHARACTER_SKILL_MACE,         5000},
+    {CHARACTER_SKILL_BLASTER,      0},
+    {CHARACTER_SKILL_SHIELD,       3000},
+    {CHARACTER_SKILL_LEATHER,      3000},
+    {CHARACTER_SKILL_CHAIN,        3000},
+    {CHARACTER_SKILL_PLATE,        3000},
+    {CHARACTER_SKILL_FIRE,         4000},
+    {CHARACTER_SKILL_AIR,          4000},
+    {CHARACTER_SKILL_WATER,        4000},
+    {CHARACTER_SKILL_EARTH,        4000},
+    {CHARACTER_SKILL_SPIRIT,       4000},
+    {CHARACTER_SKILL_MIND,         4000},
+    {CHARACTER_SKILL_BODY,         4000},
+    {CHARACTER_SKILL_LIGHT,        5000},
+    {CHARACTER_SKILL_DARK,         5000},
+    {CHARACTER_SKILL_ITEM_ID,      2500},
+    {CHARACTER_SKILL_MERCHANT,     5000},
+    {CHARACTER_SKILL_REPAIR,       2500},
+    {CHARACTER_SKILL_BODYBUILDING, 2500},
+    {CHARACTER_SKILL_MEDITATION,   2500},
+    {CHARACTER_SKILL_PERCEPTION,   2500},
+    {CHARACTER_SKILL_DIPLOMACY,    0}, // not used
+    {CHARACTER_SKILL_THIEVERY,     0}, // not used
+    {CHARACTER_SKILL_TRAP_DISARM,  2500},
+    {CHARACTER_SKILL_DODGE,        5000},
+    {CHARACTER_SKILL_UNARMED,      5000},
+    {CHARACTER_SKILL_MONSTER_ID,   2500},
+    {CHARACTER_SKILL_ARMSMASTER,   5000},
+    {CHARACTER_SKILL_STEALING,     2500},
+    {CHARACTER_SKILL_ALCHEMY,      2500},
+    {CHARACTER_SKILL_LEARNING,     5000},
+    {CHARACTER_SKILL_CLUB,         2500},
+    {CHARACTER_SKILL_MISC,         0} // hidden, not used
+}};
+
+IndexedArray<int, CHARACTER_SKILL_FIRST, CHARACTER_SKILL_LAST> grandmasterSkillMasteryCost = {{
+    {CHARACTER_SKILL_STAFF,        8000},
+    {CHARACTER_SKILL_SWORD,        8000},
+    {CHARACTER_SKILL_DAGGER,       8000},
+    {CHARACTER_SKILL_AXE,          8000},
+    {CHARACTER_SKILL_SPEAR,        8000},
+    {CHARACTER_SKILL_BOW,          8000},
+    {CHARACTER_SKILL_MACE,         8000},
+    {CHARACTER_SKILL_BLASTER,      0},
+    {CHARACTER_SKILL_SHIELD,       7000},
+    {CHARACTER_SKILL_LEATHER,      7000},
+    {CHARACTER_SKILL_CHAIN,        7000},
+    {CHARACTER_SKILL_PLATE,        7000},
+    {CHARACTER_SKILL_FIRE,         8000},
+    {CHARACTER_SKILL_AIR,          8000},
+    {CHARACTER_SKILL_WATER,        8000},
+    {CHARACTER_SKILL_EARTH,        8000},
+    {CHARACTER_SKILL_SPIRIT,       8000},
+    {CHARACTER_SKILL_MIND,         8000},
+    {CHARACTER_SKILL_BODY,         8000},
+    {CHARACTER_SKILL_LIGHT,        8000},
+    {CHARACTER_SKILL_DARK,         8000},
+    {CHARACTER_SKILL_ITEM_ID,      6000},
+    {CHARACTER_SKILL_MERCHANT,     8000},
+    {CHARACTER_SKILL_REPAIR,       6000},
+    {CHARACTER_SKILL_BODYBUILDING, 6000},
+    {CHARACTER_SKILL_MEDITATION,   6000},
+    {CHARACTER_SKILL_PERCEPTION,   6000},
+    {CHARACTER_SKILL_DIPLOMACY,    0}, // not used
+    {CHARACTER_SKILL_THIEVERY,     0}, // not used
+    {CHARACTER_SKILL_TRAP_DISARM,  6000},
+    {CHARACTER_SKILL_DODGE,        8000},
+    {CHARACTER_SKILL_UNARMED,      8000},
+    {CHARACTER_SKILL_MONSTER_ID,   6000},
+    {CHARACTER_SKILL_ARMSMASTER,   8000},
+    {CHARACTER_SKILL_STEALING,     6000},
+    {CHARACTER_SKILL_ALCHEMY,      6000},
+    {CHARACTER_SKILL_LEARNING,     8000},
+    {CHARACTER_SKILL_CLUB,         6000},
+    {CHARACTER_SKILL_MISC,         0} // hidden, not used
+}};
+
+std::array<std::pair<int16_t, ITEM_TYPE>, 27> _4F0882_evt_VAR_PlayerItemInHands_vals = {{
+    {0x0D4, ITEM_QUEST_VASE},
+    {0x0D5, ITEM_RARE_LADY_CARMINES_DAGGER},
+    {0x0D6, ITEM_MESSAGE_SCROLL_OF_WAVES},
+    {0x0D7, ITEM_MESSAGE_CIPHER},
+    {0x0D8, ITEM_QUEST_WORN_BELT},
+    {0x0D9, ITEM_QUEST_HEART_OF_THE_WOOD},
+    {0x0DA, ITEM_MESSAGE_MAP_TO_EVENMORN_ISLAND},
+    {0x0DB, ITEM_QUEST_GOLEM_HEAD},
+    {0x0DC, ITEM_QUEST_ABBEY_NORMAL_GOLEM_HEAD},
+    {0x0DD, ITEM_QUEST_GOLEM_RIGHT_ARM},
+    {0x0DE, ITEM_QUEST_GOLEM_LEFT_ARM},
+    {0x0DF, ITEM_QUEST_GOLEM_RIGHT_LEG},
+    {0x0E0, ITEM_QUEST_GOLEM_LEFT_LEG},
+    {0x0E1, ITEM_QUEST_GOLEM_CHEST},
+    {0x0E2, ITEM_SPELLBOOK_DIVINE_INTERVENTION},
+    {0x0E3, ITEM_QUEST_DRAGON_EGG},
+    {0x0E4, ITEM_QUEST_ZOKARR_IVS_SKULL},
+    {0x0E5, ITEM_QUEST_LICH_JAR_EMPTY},
+    {0x0E6, ITEM_QUEST_ELIXIR},
+    {0x0E7, ITEM_QUEST_CASE_OF_SOUL_JARS},
+    {0x0E8, ITEM_QUEST_ALTAR_PIECE_1},
+    {0x0E9, ITEM_QUEST_ALTAR_PIECE_2},
+    {0x0EA, ITEM_QUEST_CONTROL_CUBE},
+    {0x0EB, ITEM_QUEST_WETSUIT},
+    {0x0EC, ITEM_QUEST_OSCILLATION_OVERTHRUSTER},
+    {0x0ED, ITEM_QUEST_LICH_JAR_FULL},
+    {0x0F1, ITEM_RARE_THE_PERFECT_BOW}
+}};
 
 IndexedArray<int, BUILDING_WEAPON_SHOP, BUILDING_DARK_GUILD> itemAmountInShop = {{
     {BUILDING_WEAPON_SHOP,   6},
@@ -605,6 +781,246 @@ void playHouseSound(HOUSE_ID houseID, HouseSoundType type) {
         int roomSoundId = pAnimatedRooms[buildingTable[houseID].uAnimationID].uRoomSoundId;
         SoundID soundId = SoundID(std::to_underlying(type) + 100 * (roomSoundId + 300));
         pAudioPlayer->playHouseSound(soundId, true);
+    }
+}
+
+//----- (004B3EF0) --------------------------------------------------------
+void DrawJoinGuildWindow(GUILD_ID guild_id) {
+    uDialogueType = DIALOGUE_81_join_guild;
+    current_npc_text = pNPCTopics[guild_id + 99].pText;
+    GetJoinGuildDialogueOption(guild_id);
+    pDialogueWindow->Release();
+    pDialogueWindow = new GUIWindow(WINDOW_Dialogue, {0, 0}, {render->GetRenderDimensions().w, 350}, guild_id);
+    pBtn_ExitCancel = pDialogueWindow->CreateButton({471, 445}, {169, 35}, 1, 0, UIMSG_Escape, 0, Io::InputAction::Invalid,
+        localization->GetString(LSTR_CANCEL), { ui_exit_cancel_button_background }
+    );
+    pDialogueWindow->CreateButton({0, 0}, {0, 0}, 1, 0, UIMSG_HouseScreenClick, 0, Io::InputAction::Invalid, "");
+    pDialogueWindow->CreateButton({480, 160}, {140, 30}, 1, 0, UIMSG_ClickNPCTopic, DIALOGUE_82_join_guild, Io::InputAction::Invalid,
+        localization->GetString(LSTR_JOIN));
+    pDialogueWindow->_41D08F_set_keyboard_control_group(1, 1, 0, 2);
+    dialog_menu_id = DIALOGUE_OTHER;
+}
+
+void _4B3FE5_training_dialogue(int eventId) {
+    uDialogueType = DIALOGUE_SKILL_TRAINER;
+    current_npc_text = std::string(pNPCTopics[eventId + 168].pText);
+    _4B254D_SkillMasteryTeacher(eventId);  // checks whether the facility can be used
+    pDialogueWindow->Release();
+    pDialogueWindow = new GUIWindow(WINDOW_Dialogue, {0, 0}, {render->GetRenderDimensions().w, 350}, eventId);
+    pBtn_ExitCancel = pDialogueWindow->CreateButton({471, 445}, {169, 35}, 1, 0, UIMSG_Escape, 0, Io::InputAction::Invalid,
+        localization->GetString(LSTR_CANCEL), { ui_exit_cancel_button_background }
+    );
+    pDialogueWindow->CreateButton({0, 0}, {0, 0}, 1, 0, UIMSG_HouseScreenClick, 0, Io::InputAction::Invalid, "");
+    pDialogueWindow->CreateButton({480, 160}, {0x8Cu, 0x1Eu}, 1, 0, UIMSG_ClickNPCTopic, DIALOGUE_79_mastery_teacher, Io::InputAction::Invalid,
+        guild_membership_approved ? localization->GetString(LSTR_LEARN) : "");
+    pDialogueWindow->_41D08F_set_keyboard_control_group(1, 1, 0, 2);
+    dialog_menu_id = DIALOGUE_OTHER;
+}
+
+/**
+ * @offset 0x004B1ECE.
+ *
+ * @brief Oracle's 'I lost it!' dialog option
+ */
+void OracleDialogue() {
+    ItemGen *item = nullptr;
+    ITEM_TYPE item_id = ITEM_NULL;
+
+    // display "You never had it" if nothing missing will be found
+    current_npc_text = pNPCTopics[667].pText;
+
+    // only items with special subquest in range 212-237 and also 241 are recoverable
+    for (auto pair : _4F0882_evt_VAR_PlayerItemInHands_vals) {
+        int quest_id = pair.first;
+        if (pParty->_questBits[quest_id]) {
+            ITEM_TYPE search_item_id = pair.second;
+            if (!pParty->hasItem(search_item_id) && pParty->pPickedItem.uItemID != search_item_id) {
+                item_id = search_item_id;
+                break;
+            }
+        }
+    }
+
+    // missing item found
+    if (item_id != ITEM_NULL) {
+        pParty->pCharacters[0].AddVariable(VAR_PlayerItemInHands, std::to_underlying(item_id));
+        // TODO(captainurist): what if fmt throws?
+        current_npc_text = fmt::sprintf(pNPCTopics[666].pText, // "Here's %s that you lost. Be careful"
+                                        fmt::format("{::}{}\f00000", colorTable.Jonquil.tag(),
+                                                    pItemTable->pItems[item_id].pUnidentifiedName));
+    }
+
+    // missing item is lich jar and we need to bind soul vessel to lich class character
+    // TODO(Nik-RE-dev): this code is walking only through inventory, but item was added to hand, so it will not bind new item if it was acquired
+    //                   rather this code will bind jars that already present in inventory to liches that currently do not have binded jars
+    if (item_id == ITEM_QUEST_LICH_JAR_FULL) {
+        for (int i = 0; i < pParty->pCharacters.size(); i++) {
+            if (pParty->pCharacters[i].classType == CHARACTER_CLASS_LICH) {
+                bool have_vessels_soul = false;
+                for (Character &player : pParty->pCharacters) {
+                    for (int idx = 0; idx < Character::INVENTORY_SLOT_COUNT; idx++) {
+                        if (player.pInventoryItemList[idx].uItemID == ITEM_QUEST_LICH_JAR_FULL) {
+                            if (player.pInventoryItemList[idx].uHolderPlayer == -1) {
+                                item = &player.pInventoryItemList[idx];
+                            }
+                            if (player.pInventoryItemList[idx].uHolderPlayer == i) {
+                                have_vessels_soul = true;
+                            }
+                        }
+                    }
+                }
+
+                if (item && !have_vessels_soul) {
+                    item->uHolderPlayer = i;
+                    break;
+                }
+            }
+        }
+    }
+}
+
+//----- (004B254D) --------------------------------------------------------
+std::string _4B254D_SkillMasteryTeacher(int trainerInfo) {
+    uint8_t teacherLevel = (trainerInfo - 200) % 3;
+    CharacterSkillType skillBeingTaught = static_cast<CharacterSkillType>((trainerInfo - 200) / 3);
+    Character *activePlayer = &pParty->activeCharacter();
+    CharacterClassType pClassType = activePlayer->classType;
+    CharacterSkillMastery currClassMaxMastery = skillMaxMasteryPerClass[pClassType][skillBeingTaught];
+    CharacterSkillMastery masteryLevelBeingTaught = dword_F8B1B0_MasteryBeingTaught = static_cast<CharacterSkillMastery>(teacherLevel + 2);
+    guild_membership_approved = false;
+
+    if (currClassMaxMastery < masteryLevelBeingTaught) {
+        if (skillMaxMasteryPerClass[getTier2Class(pClassType)][skillBeingTaught] >= masteryLevelBeingTaught) {
+            return localization->FormatString(LSTR_FMT_HAVE_TO_BE_PROMOTED, localization->GetClassName(getTier2Class(pClassType)));
+        } else if (skillMaxMasteryPerClass[getTier3LightClass(pClassType)][skillBeingTaught] >= masteryLevelBeingTaught &&
+                skillMaxMasteryPerClass[getTier3DarkClass(pClassType)][skillBeingTaught] >= masteryLevelBeingTaught) {
+            return localization->FormatString(LSTR_FMT_HAVE_TO_BE_PROMOTED_2,
+                    localization->GetClassName(getTier3LightClass(pClassType)),
+                    localization->GetClassName(getTier3DarkClass(pClassType)));
+        } else if (skillMaxMasteryPerClass[getTier3LightClass(pClassType)][skillBeingTaught] >= masteryLevelBeingTaught) {
+            return localization->FormatString(LSTR_FMT_HAVE_TO_BE_PROMOTED, localization->GetClassName(getTier3LightClass(pClassType)));
+        } else if (skillMaxMasteryPerClass[getTier3DarkClass(pClassType)][skillBeingTaught] >= masteryLevelBeingTaught) {
+            return localization->FormatString(LSTR_FMT_HAVE_TO_BE_PROMOTED, localization->GetClassName(getTier3DarkClass(pClassType)));
+        } else {
+            return localization->FormatString(LSTR_FMT_SKILL_CANT_BE_LEARNED, localization->GetClassName(pClassType));
+        }
+    }
+
+    // Not in your condition!
+    if (!activePlayer->CanAct()) {
+        return std::string(pNPCTopics[122].pText);
+    }
+
+    // You must know the skill before you can become an expert in it!
+    int skillLevel = activePlayer->getSkillValue(skillBeingTaught).level();
+    if (!skillLevel) {
+        return std::string(pNPCTopics[131].pText);
+    }
+
+    // You are already have this mastery in this skill.
+    CharacterSkillMastery skillMastery = activePlayer->getSkillValue(skillBeingTaught).mastery();
+    if (std::to_underlying(skillMastery) > teacherLevel + 1) {
+        return std::string(pNPCTopics[teacherLevel + 128].pText);
+    }
+
+    dword_F8B1AC_skill_being_taught = skillBeingTaught;
+
+    bool canLearn = true;
+
+    if (masteryLevelBeingTaught == CHARACTER_SKILL_MASTERY_EXPERT) {
+        canLearn = skillLevel >= 4;
+        gold_transaction_amount = expertSkillMasteryCost[skillBeingTaught];
+    }
+
+    if (masteryLevelBeingTaught == CHARACTER_SKILL_MASTERY_MASTER) {
+        switch (skillBeingTaught) {
+          case CHARACTER_SKILL_LIGHT:
+            canLearn = pParty->_questBits[114];
+            break;
+          case CHARACTER_SKILL_DARK:
+            canLearn = pParty->_questBits[110];
+            break;
+          case CHARACTER_SKILL_MERCHANT:
+            canLearn = activePlayer->GetBasePersonality() >= 50;
+            break;
+          case CHARACTER_SKILL_BODYBUILDING:
+            canLearn = activePlayer->GetBaseEndurance() >= 50;
+            break;
+          case CHARACTER_SKILL_LEARNING:
+            canLearn = activePlayer->GetBaseIntelligence() >= 50;
+            break;
+          default:
+            break;
+        }
+        canLearn = canLearn && (skillLevel >= 7);
+        gold_transaction_amount = masterSkillMasteryCost[skillBeingTaught];
+    }
+
+    if (masteryLevelBeingTaught == CHARACTER_SKILL_MASTERY_GRANDMASTER) {
+        switch (skillBeingTaught) {
+          case CHARACTER_SKILL_LIGHT:
+            canLearn = activePlayer->isClass(CHARACTER_CLASS_ARCHAMGE) || activePlayer->isClass(CHARACTER_CLASS_PRIEST_OF_SUN);
+            break;
+          case CHARACTER_SKILL_DARK:
+            canLearn = activePlayer->isClass(CHARACTER_CLASS_LICH) || activePlayer->isClass(CHARACTER_CLASS_PRIEST_OF_MOON);
+            break;
+          case CHARACTER_SKILL_DODGE:
+            canLearn = activePlayer->pActiveSkills[CHARACTER_SKILL_UNARMED].level() >= 10;
+            break;
+          case CHARACTER_SKILL_UNARMED:
+            canLearn = activePlayer->pActiveSkills[CHARACTER_SKILL_DODGE].level() >= 10;
+            break;
+          default:
+            break;
+        }
+        canLearn = canLearn && (skillLevel >= 10);
+        gold_transaction_amount = grandmasterSkillMasteryCost[skillBeingTaught];
+    }
+
+    // You don't meet the requirements, and cannot be taught until you do.
+    if (!canLearn) {
+        return std::string(pNPCTopics[127].pText);
+    }
+
+    // You don't have enough gold!
+    if (gold_transaction_amount > pParty->GetGold()) {
+        return std::string(pNPCTopics[124].pText);
+    }
+
+    guild_membership_approved = true;
+
+    return localization->FormatString(
+        LSTR_FMT_BECOME_S_IN_S_FOR_D_GOLD,
+        localization->MasteryNameLong(masteryLevelBeingTaught),
+        localization->GetSkillName(skillBeingTaught),
+        gold_transaction_amount
+    );
+}
+
+//----- (004B29F2) --------------------------------------------------------
+const std::string &GetJoinGuildDialogueOption(GUILD_ID guild_id) {
+    static const int dialogue_base = 110;
+    guild_membership_approved = false;
+    dword_F8B1AC_award_bit_number = static_cast<AwardType>(Award_Membership_ElementalGuilds + std::to_underlying(guild_id));
+    gold_transaction_amount = priceForMembership[guild_id];
+
+    // TODO(pskelton): check this behaviour
+    if (!pParty->hasActiveCharacter())
+        pParty->setActiveToFirstCanAct();  // avoid nzi
+
+    if (pParty->activeCharacter().CanAct()) {
+        if (pParty->activeCharacter()._achievedAwardsBits[dword_F8B1AC_award_bit_number]) {
+            return pNPCTopics[dialogue_base + 13].pText;
+        } else {
+            if (gold_transaction_amount <= pParty->GetGold()) {
+                guild_membership_approved = true;
+                return pNPCTopics[dialogue_base + guild_id].pText;
+            } else {
+                return pNPCTopics[dialogue_base + 14].pText;
+            }
+        }
+    } else {
+        return pNPCTopics[dialogue_base + 12].pText;
     }
 }
